# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Job Vacancy API
  version: 0.6.0
  description: |
    Shared contract for the web (React) and mobile (Flutter) clients.

    **Realtime.** POST /ws/ticket, then open `wss://…/api/v1/ws?ticket=…`. Server events:
    `ready`, `message.new`, `message.read`, `message.deleted`, `typing`, `presence`,
    `presence.snapshot`, `notification.new`, `pong`, `error` — each `{"type", "data"}`.
    Client frames: `{"type":"ping"}`, `{"type":"typing","conversation_id"}`,
    `{"type":"read","conversation_id","message_id"}`, `{"type":"watch_presence","user_ids":[…]}`.
    Events are not buffered: after reconnecting, fetch `/conversations/{id}/messages?after=<last id>`.

    **Uploads.** POST /files → multipart POST to `upload.url` with every `upload.fields`
    entry followed by the `file` field → POST /files/{id}/complete → use the file id.

    **Envelope.** Success: `{"data": ...}` (lists add `"meta"`). Failure:
    `{"error": {"code", "message", "fields"?, "retry_after"?}}`. `code` is stable; clients
    translate it into uz / uz-Cyrl / ru / en. `fields` maps a JSON field to the failing
    rule (`required`, `email`, `min=8`, `password`, `oneof=...`).

    **Deadlines.** Every request runs under a server deadline (8 s; PDF export 25 s;
    the WebSocket is exempt) and each database statement under 5 s. When one is exceeded
    the API answers `503` with code `timeout` and `Retry-After: 1`; the call is safe to
...
    limits of their own.

...

...
    it rendered.

...

...
servers:
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    CompanyRef: { name: company, in: path, required: true, schema: { type: string }, description: uuid or slug }
    IfNoneMatch:
      name: If-None-Match
      in: header
      schema: { type: string }
      description: ETag of a copy the client has; unchanged content answers 304 with no body.
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, examples: [invalid_credentials] }
            message: { type: string }
            fields: { type: object, additionalProperties: { type: string } }
            retry_after: { type: integer, description: seconds }
...
    CompanyInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        industry_id: { type: integer, description: category id }
        size: { type: string, enum: ["1-10", "11-50", "51-200", "201-500", "501-1000", "1000+"] }
        website: { type: string, description: "scheme optional: najottalim.uz → https://najottalim.uz" }
        email: { type: string, format: email }
        phone: { type: string }
        region_id: { type: integer, description: top-level region }
        address: { type: string, maxLength: 300 }
        about: { type: string, maxLength: 5000 }
        founded_year: { type: integer, minimum: 1800, maximum: 2100 }
    Company:
      allOf:
        - $ref: "#/components/schemas/CompanyInput"
        - type: object
          required: [id, slug, name, logo_url, cover_url, verified, created_at, about]
          properties:
            id: { type: string, format: uuid }
            slug: { type: string }
            logo_url: { type: [string, "null"] }
            cover_url: { type: [string, "null"] }
            verified: { type: boolean }
            created_at: { type: string, format: date-time }
            open_vacancies: { type: integer, description: public profile only }
            my_role: { type: string, enum: [owner, admin, recruiter], description: "/me/companies only" }
            status: { type: string, enum: [active, blocked, deleted], description: "/me/companies only; blocked companies can't be managed" }
...
    CompanyInvite:
      type: object
      required: [id, email, role, invited_by_name, expires_at, expired, created_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { type: string, enum: [admin, recruiter] }
        invited_by_name: { type: [string, "null"] }
        expires_at: { type: string, format: date-time }
        expired: { type: boolean, description: "past expires_at: resend it with POST /companies/{company}/invites" }
        created_at: { type: string, format: date-time }
...
  headers:
    ETag: { description: Validator of the response body (weak), for If-None-Match, schema: { type: string } }
...
  responses:
    NotModified:
      description: Not modified (the If-None-Match ETag is current); no body
      headers:
        ETag: { $ref: "#/components/headers/ETag" }
    Error:
      description: Error
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    NoContent: { description: Done }
...
    Company:
      description: Company
      content:
        application/json:
          schema: { type: object, properties: { data: { $ref: "#/components/schemas/Company" } } }
paths:
...
  /companies:
    get:
      tags: [companies]
      summary: Company directory (verified first, then by open vacancies), 24 per page
      description: |
        Keyset pages (TZ BE-03), never OFFSET. Page either by number (`page`, for numbered
        pagination: it seeks from a cached page anchor) or by `cursor` (from
        `meta.next_cursor`, takes precedence over `page`). `meta.total` and
        `meta.page_count` come from page anchors cached for ~3 minutes, so they (and page
        boundaries) may lag new companies or ranking changes by that long. A `page` past
        the last one returns an empty `data`. `open_vacancies` is kept by a database
        trigger on every vacancy status change.
      parameters:
        - { name: q, in: query, description: "Name contains (case-insensitive, max 100 chars)", schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: cursor, in: query, description: "meta.next_cursor of the previous page", schema: { type: string } }
      responses:
        "200":
          description: Page
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Company" } }
                  meta:
                    type: object
                    required: [page, next_page, next_cursor, total, page_count]
                    properties:
                      page: { type: [integer, "null"], description: "Requested page number; null when paging by cursor" }
                      next_page: { type: [integer, "null"] }
                      next_cursor: { type: [string, "null"] }
                      total: { type: integer, description: "Matching active companies" }
                      page_count: { type: integer }
        "400": { $ref: "#/components/responses/Error" }
    post:
      tags: [companies]
      security: [{ bearer: [] }]
      summary: Create a company (employer with a verified email or phone); caller becomes owner
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CompanyInput" } } } }
      responses:
        "201": { $ref: "#/components/responses/Company" }
        "403": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /companies/{company}:
    parameters: [{ $ref: "#/components/parameters/CompanyRef" }]
    get:
      tags: [companies]
      summary: Public company profile
      description: Cached (~60 s, dropped on every profile, logo, verification or vacancy status change); ETag, 304.
      parameters: [{ $ref: "#/components/parameters/IfNoneMatch" }]
      responses:
        "200": { $ref: "#/components/responses/Company" }
        "304": { $ref: "#/components/responses/NotModified" }
        "404": { $ref: "#/components/responses/Error" }
    put:
      tags: [companies]
      security: [{ bearer: [] }]
      summary: Replace the profile (owner/admin). The slug never changes.
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CompanyInput" } } } }
      responses:
        "200": { $ref: "#/components/responses/Company" }
        "403": { $ref: "#/components/responses/Error" }
  /companies/{company}/members:
    parameters: [{ $ref: "#/components/parameters/CompanyRef" }]
    get:
      tags: [companies]
      security: [{ bearer: [] }]
      summary: Team members (any member)
      responses: { "200": { description: Members } }
    post:
      tags: [companies]
      security: [{ bearer: [] }]
      deprecated: true
      summary: "Invite by e-mail, or change the role of someone already on the team (use /invites and PATCH /members/{user})"
      description: |
        Kept for compatibility (TZ FN-05): nobody is added directly any more. For a current
        member it changes the role; for anyone else it sends an invite (same rules as
        POST /companies/{company}/invites) that they have to accept.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties: { email: { type: string, format: email }, role: { type: string, enum: [admin, recruiter] } }
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /companies/{company}/members/{user}:
    delete:
      tags: [companies]
      security: [{ bearer: [] }]
      summary: Remove a member (owner/admin) or leave the company (yourself)
      parameters:
        - { $ref: "#/components/parameters/CompanyRef" }
        - { name: user, in: path, required: true, schema: { type: string, format: uuid } }
      responses: { "204": { $ref: "#/components/responses/NoContent" } }
    patch:
      tags: [companies]
      security: [{ bearer: [] }]
      summary: Change a member's role (admins manage recruiters; the owner also admins)
      parameters:
        - { $ref: "#/components/parameters/CompanyRef" }
        - { name: user, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [role], properties: { role: { type: string, enum: [admin, recruiter] } } } } }
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /companies/{company}/invites:
    parameters: [{ $ref: "#/components/parameters/CompanyRef" }]
    get:
      tags: [companies]
      security: [{ bearer: [] }]
      summary: Pending invites (owner/admin), expired ones flagged
      responses:
        "200":
          description: Invites
          content: { application/json: { schema: { type: object, properties: { data: { type: array, items: { $ref: "#/components/schemas/CompanyInvite" } } } } } }
        "403": { $ref: "#/components/responses/Error" }
    post:
      tags: [companies]
      security: [{ bearer: [] }]
      summary: Invite someone by e-mail (TZ FN-05; valid 7 days, resending refreshes it)
      description: |
        The invitee is told in the app and by e-mail when they have an account, by e-mail
        otherwise (a link to /me/invites). They become a member only after accepting; until
        then they can't read the company's team, vacancies or applications. Admins invite
        recruiters; the owner also invites admins. Errors: 409 already_member,
        404 member_not_found (the address belongs to a non-employer or blocked account),
        429 rate_limited (a company sends at most 50 invites a day).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties: { email: { type: string, format: email }, role: { type: string, enum: [admin, recruiter] } }
      responses:
        "201":
          description: Invite
          content: { application/json: { schema: { type: object, properties: { data: { $ref: "#/components/schemas/CompanyInvite" } } } } }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /companies/{company}/invites/{invite}:
    delete:
      tags: [companies]
      security: [{ bearer: [] }]
      summary: Revoke a pending invite (owner/admin)
      parameters:
        - { $ref: "#/components/parameters/CompanyRef" }
        - { name: invite, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        "404": { $ref: "#/components/responses/Error" }
  /companies/{company}/owner:
    post:
      tags: [companies]
      security: [{ bearer: [] }]
      summary: Hand the company to another member (owner only); the old owner stays as admin
      parameters: [{ $ref: "#/components/parameters/CompanyRef" }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [user_id], properties: { user_id: { type: string, format: uuid } } } } }
      responses:
        "200": { $ref: "#/components/responses/Company" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
  /companies/{company}/logo:
    put:
      tags: [companies]
      security: [{ bearer: [] }]
      summary: Set or remove the company logo (owner/admin)
      parameters: [{ $ref: "#/components/parameters/CompanyRef" }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { file_id: { type: [string, "null"], format: uuid } } } } } }
      responses: { "200": { $ref: "#/components/responses/Company" } }
...